requirements.txt
streamlit
pandas
google-analytics-data
google-analytics-admin
gspread
google-generativeai
google-auth